changes from jari

+++ app/src/main/java/com/teamx/equiz/data/models/orderDetailData/ProductDetail.kt
@@ -3,6 +3,6 @@ package com.teamx.equiz.data.models.orderDetailData
 data class ProductDetail(
     val _id: String,
     val product: Product,
-    val quantity: Int,
-    val totalPoint: Int
+    val quantity: Double,
+    val totalPoint: Double
 )
